requirements.txt
snap-stanford
matplotlib
numpy
pandas
scipy
